refactor(pipeline): relocate reusable types out of unified_pipeline (R6a)
Repoint the surviving importers (fastq_parse, grouper, mi_group, template,
commands/common) at fgumi_bam_io directly for DecodedRecord, GroupKey,
GroupKeyConfig, Grouper, and MemoryEstimate, which already live there and
were only reached through a unified_pipeline re-export.

Relocate the three items still actually defined in unified_pipeline that
surviving code needs: the BatchWeight trait moves to a new root-crate
module (src/lib/batch_weight.rs) rather than fgumi-bam-io, since its
surviving impls target foreign types and would violate the orphan rule
there; SchedulerStrategy moves to commands/common.rs, next to its only
surviving consumer (SchedulerOptions); the backpressure constants and
stage_high_water_mark move to a new src/lib/pipeline/backpressure.rs.
unified_pipeline keeps pub use shims for all three so the still-present
legacy command paths keep compiling unchanged across every feature
configuration.

Repoint the two grouper unit tests that called the 4-arg
unified_pipeline::compute_group_key_from_raw to the canonical 3-arg
fgumi_bam_io version, dropping the discarded UMI offset return value the
tests never used; the resulting GroupKey is unchanged for valid records.

BamPipelineConfig stays legacy-only and is left in place.
