+++ src/lib/fastq_parse.rs
@@ -6,7 +6,7 @@
 
 use std::io;
 
-use crate::unified_pipeline::MemoryEstimate;
+use fgumi_bam_io::MemoryEstimate;
 
 /// A parsed FASTQ record stored as a single heap allocation.
 ///
